requirements.txt
openai
python-frontmatter